% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_value_extract.R
\name{dic_value_extract_one_dataset}
\alias{dic_value_extract_one_dataset}
\title{extract information from one single dataset provided as data frame}
\usage{
dic_value_extract_one_dataset(.data)
}
\arguments{
\item{.data}{the dataframe from which key values
will be extracted}
}
\value{
a dataframe including the values extracted from each variable in the dataset
}
\description{
extract information from one single dataset provided as data frame
}
\examples{

require(haven)
# mtcars.sas7bdat was generated from R mtcars dataset
df <- system.file("extdata", "mtcars.sas7bdat", package = "createdictionary")
df \%>\%
read_sas() \%>\%
  dic_value_extract_one_dataset()
# should work if there exist sas7bdat datasets in R path.
# The package "haven" have such datasets.

df <- system.file(package = "haven")
dse <- flat2file(df, "*.sas7bdat")
dse[1] \%>\%
 read_sas() \%>\%
 dic_value_extract_one_dataset()

# need access to the specified folder and have sas7bdat datasets there
df <- "/LSC/NES/study/CARDIA/core datasets/Y25/DATA"
dse <- flat2file(df, "*.sas7bdat")
d <- dse[3] \%>\% read_sas()
d \%>\% dic_value_extract_one_dataset()


}
