% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_value_extract.R
\name{dic_value_extract_one_dataset_path}
\alias{dic_value_extract_one_dataset_path}
\title{extract information from one single dataset provided as file path}
\usage{
dic_value_extract_one_dataset_path(dataset_path)
}
\arguments{
\item{dataset_path}{the full path to the dataset from which key values
will be extracted}
}
\value{
a dataframe including the values extracted from each variable in the dataset
}
\description{
extract information from one single dataset provided as file path
}
\examples{

# mtcars.sas7bdat was generated from R mtcars dataset
df <- system.file("extdata", "mtcars.sas7bdat", package = "createdictionary")
df \%>\%
  dic_value_extract_one_dataset_path()
# should work if there exist sas7bdat datasets in R path.
# The package "haven" have such datasets.
df <- system.file(package = "haven")
dse <- flat2file(df, "*.sas7bdat")
dse[1] \%>\%
  dic_value_extract_one_dataset_path()

# need access to the specified folder and have sas7bdat datasets there
df <- "/LSC/NES/study/CARDIA/core datasets/Y25/DATA"
dse <- flat2file(df, "*.sas7bdat")
dse[3] \%>\%
  dic_value_extract_one_dataset_path()

value_extracted <-  dse \%>\%
  map_dfr(dic_value_extract_one_dataset_path)


}
