% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_value_extract.R
\name{dic_value_extract_one_var_safely}
\alias{dic_value_extract_one_var_safely}
\title{the safe version of extract values from one variable
require(haven)
# mtcars.sas7bdat was generated from R mtcars dataset
# get file path
df <- system.file("extdata", "mtcars.sas7bdat", package = "createdictionary")
# read in data
d <- read_sas(df)
# one variable
dic_value_extract_one_var_safetly(d$mpg, "mpg")
dic_value_extract_one_var_safetly(d$mpg, "mpg")$result}
\usage{
dic_value_extract_one_var_safely(...)
}
\description{
the safe version of extract values from one variable
require(haven)
# mtcars.sas7bdat was generated from R mtcars dataset
# get file path
df <- system.file("extdata", "mtcars.sas7bdat", package = "createdictionary")
# read in data
d <- read_sas(df)
# one variable
dic_value_extract_one_var_safetly(d$mpg, "mpg")
dic_value_extract_one_var_safetly(d$mpg, "mpg")$result
}
