% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat2file.R
\name{flat2file}
\alias{flat2file}
\title{Concatenate all the files inside or outside directories}
\usage{
flat2file(df, pattern)
}
\arguments{
\item{df}{character vector or list of files or directories in full path}

\item{pattern}{file selection pattern}
}
\value{
character vector including full path only for files
}
\description{
Concatenate all the files inside or outside directories
}
\examples{
df <- c("/opt/R/4.0.2/lib/R/library/base/Meta/Rd.rds", system.file())
df <- c("/opt/R/4.0.2/lib/R/library/base/Meta/Rd.rds", system.file(package = "readr"), system.file(package = "dplyr"))
flat2file(df, "*.rds")
df <- "/LSC/NES/study/CARDIA/core datasets/Y25/DATA"
flat2file(df, "*.sas7bdat")


}
